fix load data category and district, view item ping job
