+++ app/src/main/java/com/example/phuong/viectimnguoiapp/activities/LoginActivity1.java
@@ -7,6 +7,8 @@ import android.widget.Toast;
 
 import com.example.phuong.viectimnguoiapp.R;
 import com.example.phuong.viectimnguoiapp.databases.RealmHelper;
+import com.example.phuong.viectimnguoiapp.objects.CategoryJob;
+import com.example.phuong.viectimnguoiapp.objects.District;
 import com.example.phuong.viectimnguoiapp.objects.User;
 import com.example.phuong.viectimnguoiapp.utils.Constant;
 import com.example.phuong.viectimnguoiapp.utils.DialogLoading;
@@ -126,6 +128,9 @@ public class LoginActivity1 extends AppCompatActivity implements GoogleApiClient
                     mapUser.put("phone", "");
                     mapUser.put("status", Constant.USER_ACTIVE);
                     mapUser.put("point", "0");
+                    mapUser.put("id",FirebaseAuth.getInstance().getCurrentUser().getUid());
+                    mapUser.put("email",FirebaseAuth.getInstance().getCurrentUser().getEmail());
+                    mapUser.put("username",FirebaseAuth.getInstance().getCurrentUser().getDisplayName());
                     mFirebaseUserInfor.push().setValue(mapUser);
                     mUser.setPhone("");
                     mUser.setAddress("");
@@ -245,9 +250,49 @@ public class LoginActivity1 extends AppCompatActivity implements GoogleApiClient
             MainActivity_.intent(LoginActivity1.this).start();
             getUserInfor();
             getSetting();
+            getDataCategoryJob();
+            getDataDistrict();
             finish();
         } else {
         }
+    }
+
+    public void getDataDistrict() {
+        DatabaseReference mFirebaseDistrict = FirebaseDatabase.getInstance().getReference("/districts");
+        mFirebaseDistrict.addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                for (DataSnapshot obj : dataSnapshot.getChildren()) {
+                    HashMap<String, Object> map = (HashMap<String, Object>) obj.getValue();
+                    District district = new District(map.get("id").toString(), map.get("name").toString());
+                    mData.addDistrict(district);
+                }
+            }
+
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+
+            }
+        });
+    }
+
+    public void getDataCategoryJob() {
+        DatabaseReference mFirebaseCategoryJob = FirebaseDatabase.getInstance().getReference("/categoryJobs");
+        mFirebaseCategoryJob.addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                for (DataSnapshot obj : dataSnapshot.getChildren()) {
+                    HashMap<String, Object> map = (HashMap<String, Object>) obj.getValue();
+                    CategoryJob categoryJob = new CategoryJob(map.get("id").toString(), map.get("name").toString());
+                    mData.addCategoryJob(categoryJob);
+                }
+            }
+
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+
+            }
+        });
     }
 
     private void requestLogOutGoogle() {
