+++ app/src/main/res/drawable/choice_user_contact.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:drawable="@drawable/ic_chose" android:state_checked="false"></item>
+    <item android:drawable="@drawable/ic_chosen" android:state_checked="true"></item>
+
+</selector>
