+++ app/src/main/res/layout/fragment_list_user_ping_by_new.xml
@@ -13,5 +13,5 @@
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recyclerViewUserPing"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"></android.support.v7.widget.RecyclerView>
+        android:layout_height="match_parent"></android.support.v7.widget.RecyclerView>
 </RelativeLayout>
